requirements.txt
joblib == 1.3.2
scikit-learn
newspaper3k

